--- beer-song/Cargo.toml
@@ -1,3 +1,0 @@
-[package]
-name = "beer-song"
-version = "0.0.0"
