--- bob/Cargo.toml
@@ -1,3 +1,0 @@
-[package]
-name = "bob"
-version = "1.2.0"
