+++ exercises/gigasecond/src/lib.rs
@@ -1,7 +1,8 @@
 extern crate chrono;
-use chrono::{DateTime, Utc};
+use chrono::{DateTime, Utc, Duration};
 
 // Returns a Utc DateTime one billion seconds after start.
 pub fn after(start: DateTime<Utc>) -> DateTime<Utc> {
-    unimplemented!("What time is a gigasecond later than {}", start);
+	let dur = Duration::seconds(1000000000);
+	start + dur
 }
