+++ exercises/leap/tests/leap.rs
@@ -6,19 +6,19 @@ fn test_vanilla_leap_year() {
 }
 
 #[test]
-#[ignore]
+
 fn test_any_old_year() {
     assert_eq!(leap::is_leap_year(1997), false);
 }
 
 #[test]
-#[ignore]
+
 fn test_century() {
     assert_eq!(leap::is_leap_year(1900), false);
 }
 
 #[test]
-#[ignore]
+
 fn test_exceptional_centuries() {
     assert_eq!(leap::is_leap_year(2000), true);
     assert_eq!(leap::is_leap_year(2400), true);
