+++ exercises/nth-prime/tests/nth-prime.rs
@@ -6,25 +6,25 @@ fn test_first_prime() {
 }
 
 #[test]
-#[ignore]
+
 fn test_second_prime() {
     assert_eq!(np::nth(2), Some(3));
 }
 
 #[test]
-#[ignore]
+
 fn test_sixth_prime() {
     assert_eq!(np::nth(6), Some(13));
 }
 
 #[test]
-#[ignore]
+
 fn test_big_prime() {
     assert_eq!(np::nth(10001), Some(104743));
 }
 
 #[test]
-#[ignore]
+
 fn test_zeroth_prime() {
     assert_eq!(np::nth(0), None);
 }
