+++ exercises/reverse-string/src/lib.rs
@@ -1,3 +1,6 @@
+extern crate unicode_segmentation;
+use unicode_segmentation::UnicodeSegmentation;
+
 pub fn reverse(input: &str) -> String {
-    unimplemented!("Write a function to reverse {}", input);
+	input.graphemes(true).rev().collect()
 }
