--- gigasecond/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "gigasecond"
-version = "1.1.0"
-
-[dependencies]
-chrono = "0.4"
-
