--- hello-world/Cargo.toml
@@ -1,3 +1,0 @@
-[package]
-name = "hello-world"
-version = "1.1.0"
