--- leap/Cargo.toml
@@ -1,3 +1,0 @@
-[package]
-name = "leap"
-version = "1.0.0"
