--- leap/src/lib.rs
@@ -1,3 +1,0 @@
-pub fn is_leap_year(year: i32) -> bool {
-    year % 4 == 0 && year % 100 != 0 ||  year % 400 == 0   
-}
