--- nth-prime/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "nth_prime"
-version = "1.0.0"
-authors = ["sacherjj <sacherjj@gmail.com>"]
-
-[dependencies]
