--- raindrops/Cargo.toml
@@ -1,3 +1,0 @@
-[package]
-name = "raindrops"
-version = "1.1.0"
