--- reverse-string/.gitignore
@@ -1,3 +1,0 @@
-# Ignore Cargo.lock if creating a library
-# More information here http://doc.crates.io/guide.html#cargotoml-vs-cargolock
-Cargo.lock
