--- reverse-string/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "reverse_string"
-version = "1.0.0"
-authors = ["Peter Goodspeed-Niklaus <peter.r.goodspeedniklaus@gmail.com>"]
-
-[dependencies]
-unicode-segmentation = "1.1.0"
