--- reverse-string/src/lib.rs
@@ -1,6 +1,0 @@
-extern crate unicode_segmentation;
-use unicode_segmentation::UnicodeSegmentation;
-
-pub fn reverse(input: &str) -> String {
-	input.graphemes(true).rev().collect()
-}
